% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cat.list}
\alias{cat.list}
\title{Prints a list of vectors to individual files}
\usage{
cat.list(l, prefix = "out_", suffix = ".txt")
}
\arguments{
\item{l}{A list of vectors}

\item{prefix}{A prefix to use}

\item{suffix}{A suffix to use}
}
\value{
None
}
\description{
One file is generated per vector in the list.
Files are named {suffix}{list_element_name}{prefix}
}
\keyword{cat}
\keyword{util}

