% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileops.R
\name{catnf}
\alias{catnf}
\title{Writes to a file, elements seperated by newlines}
\usage{
catnf(vec, file)
}
\arguments{
\item{vec}{A vector}

\item{file}{Filename to write to}
}
\description{
Cats with newlines to a file
}
\keyword{util}

