% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clusterDEList}
\alias{clusterDEList}
\title{Clusters genes by expression using k-means}
\usage{
clusterDEList(df, clusters)
}
\arguments{
\item{df}{A data.frame containing DE values, one gene per line (gene as rowname)}

\item{clusters}{Number of clusters to generate}
}
\value{
A list of data.frames
}
\description{
Returns a list of data.frames split by cluster
}
\keyword{clustering}
\keyword{genes,}

