% Generated by roxygen2 (4.0.2): do not edit by hand
\name{factorize}
\alias{factorize}
\title{Converts columns to factors}
\usage{
factorize(df, cols = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{cols}{A int or chr vector of cols to factorize}
}
\value{
df A data.frame
}
\description{
Converts columns to factors. Defaults to all chr columns.
}
\keyword{data.frame}

