% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getUpstreamSeq}
\alias{getUpstreamSeq}
\title{Gets flanking regions of given genes}
\usage{
getUpstreamSeq(genes, len)
}
\arguments{
\item{genes}{A GRanges object of genes of interest (e.g. as produced by getDmelGenes)}

\item{len}{Length of upstream sequence to retrieve}
}
\value{
A DNAStringSet contining upstream sequences
}
\description{
Returns the upstream sequences given a list of genes
}
\keyword{genes}

