% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggsave_page}
\alias{ggsave_page}
\title{Performs ggsave with filename and dim 11 x 8.5
param filename The filename for saved file}
\usage{
ggsave_page(filename)
}
\description{
Performs ggsave with filename and dim 11 x 8.5
param filename The filename for saved file
}
\keyword{ggplot2}

