% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{grep.row}
\alias{grep.row}
\title{Returns a data.frame grep'd on rownames}
\usage{
grep.row(pattern, df)
}
\arguments{
\item{pattern}{Pattern to search for}

\item{df}{data.frame to grep}
}
\value{
A data.frame
}
\description{
Returns a data.frame grep'd on rownames
}
\keyword{data.frame}
\keyword{grep}

