% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grepcol}
\alias{grepcol}
\title{Returns a data.frame grep'd on colnames}
\usage{
grepcol(pattern, df)
}
\arguments{
\item{pattern}{Pattern to search for}

\item{df}{data.frame to grep}
}
\value{
A data.frame
}
\description{
Returns a data.frame grep'd on colnames
}
\keyword{data.frame}
\keyword{grep}

