% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{interleave}
\alias{interleave}
\title{Interleaves two vectors}
\usage{
interleave(a, b)
}
\arguments{
\item{a}{vector 1}

\item{b}{vector 2}
}
\value{
A vector
}
\description{
Interleaves two vectors. Vectors may be of unequal length, but should be of the same type.
}
\keyword{vector}

