% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lh}
\alias{lh}
\title{Lists all objects and sizes}
\usage{
lh(units = "b")
}
\arguments{
\item{u}{Units for display, as listed in the object.size help ("b", "Kb", "Mb", "Gb", etc)}
}
\value{
A data.frame containing a list of objects and their sizes
}
\description{
Lists all objects and sizes
}
\keyword{ls}
\keyword{utility}

