% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{loadBC}
\alias{loadBC}
\title{Load script for intstalling Bioconductor packages}
\usage{
loadBC()
}
\description{
Sources the Bioconductor script from their website.
Afterwards, use function biocLite("pkg name") to install/update pkgs.
}

