% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadPCRMeltDeriv}
\alias{loadPCRMeltDeriv}
\title{Loads the melting deriv xlsx from a qPCR run into a melted data.frame}
\usage{
loadPCRMeltDeriv(melt.file)
}
\arguments{
\item{melt.file}{the filename of the xlsx spreedsheet to load}
}
\description{
Loads the melting deriv xlsx from a qPCR run into a melted data.frame
}
\keyword{qPCR}

