% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadPCRQuant}
\alias{loadPCRQuant}
\title{Loads the quantitation xlsx from a qPCR run into a melted data.frame}
\usage{
loadPCRQuant(quant.file)
}
\arguments{
\item{quant.file}{the filename of the xlsx spreedsheet to load}
}
\description{
Loads the quantitation xlsx from a qPCR run into a melted data.frame
}
\keyword{qPCR}

