% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{logd}
\alias{logd}
\title{Sends timestamped log message to STDERR}
\usage{
logd(..., timestamp = T)
}
\arguments{
\item{timestamp}{Whether or not to include timestamp}
}
\description{
Sends timestamped log message to STDERR
}
\keyword{log}
\keyword{utility}

