% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileops.R
\name{multMergeCSV}
\alias{multMergeCSV}
\title{Merge all csv files in a path into one data frame}
\usage{
multMergeCSV(mypath)
}
\arguments{
\item{mypath}{the path containing files to load (all files will be loaded)}
}
\description{
This function opens all files in the given directory and merges
them into one data. Files must meet the following requirement:\
     - Merge value in first column
     - Must have headers
     - Header value for data to be merged must be unique
}
\keyword{merge}

