% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clipboard.R
\name{pbhist}
\alias{pbhist}
\title{Copies history to clipboard}
\usage{
pbhist(n = 1000)
}
\arguments{
\item{n}{Number of lines to print}
}
\description{
Copies history to clipboard
}
\keyword{clipboard}

