% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileops.R
\name{read_table_as_dt}
\alias{read_table_as_dt}
\title{Returns a data_table using read.table}
\usage{
read_table_as_dt(...)
}
\value{
A data_table
}
\description{
Returns a data_table using read.table
}
\keyword{data_table}
\keyword{write.table}

