% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataframe.R
\name{rowtofirst}
\alias{rowtofirst}
\title{Removes a data.frame's rownames and adds them as the first column}
\usage{
rowtofirst(df, name = "Row")
}
\arguments{
\item{df}{A data.frame}

\item{name}{Col name for added column}
}
\value{
A data.frame
}
\description{
Removes a data.frame's rownames and adds them as the first column
}
\keyword{data.frame}

