% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarizeqPCR}
\alias{summarizeqPCR}
\title{Saves several pdfs summarizing qPCR data}
\usage{
summarizeqPCR(prefix, quant.file, melt.file)
}
\arguments{
\item{prefix}{Prefix for files}

\item{quant.file}{the filename of the xlsx spreedsheet to load}

\item{melt.file}{the filename of the xlsx spreedsheet to load}
}
\description{
Saves several pdfs summarizing qPCR data
}
\keyword{qPCR}

