% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fileops.R
\name{write.tsv}
\alias{write.tsv}
\title{Writes a TSV file}
\usage{
write.tsv(x, file, rownames = F, ...)
}
\arguments{
\item{x}{Data.frame printable object}

\item{file}{Filename to write to}

\item{rownames}{Print rownames? Defaults to FALSE}
}
\value{
None
}
\description{
Writes a TSV file using write.table. Defaults to:
   Tab for sep
   No quoting
   No row.names
}
\keyword{IO}
\keyword{file}

