% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataframe.R
\name{factorize}
\alias{factorize}
\title{Converts columns to factors}
\usage{
factorize(df, cols = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{cols}{A int or chr vector of cols to factorize}
}
\value{
df A data.frame
}
\description{
Converts columns to factors. Defaults to all chr columns.
}
\keyword{data.frame}

