% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataframe.R
\name{grep.col}
\alias{grep.col}
\title{Returns a data.frame grep'd on colnames}
\usage{
grep.col(pattern, df)
}
\arguments{
\item{pattern}{Pattern to search for}

\item{df}{data.frame to grep}
}
\value{
A data.frame
}
\description{
Returns a data.frame grep'd on colnames
}
\keyword{data.frame}
\keyword{grep}

