% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{lh}
\alias{lh}
\title{Lists all objects and sizes}
\usage{
lh(units = "b")
}
\arguments{
\item{u}{Units for display, as listed in the object.size help ("b", "xKb", "Mb", "Gb", etc)}
}
\value{
A data.frame containing a list of objects and their sizes
}
\description{
Lists all objects and sizes
}
\keyword{ls}
\keyword{utility}

