% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileops.R
\name{multMerge}
\alias{multMerge}
\title{Merge all tsv files in a path into one data frame}
\usage{
multMerge(mypath, header = T)
}
\arguments{
\item{mypath}{the path containing files to load (all files will be loaded)}

\item{header}{do files have a header? If not, generate based on filenames}
}
\description{
This function opens all files in the given directory and merges
them into one dataframe. Files must meet the following requirement:\
     - Merge value in first column
     - Must have headers
     - Header value for data to be merged must be unique
}
\keyword{merge}

