% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multMerge}
\alias{multMerge}
\title{Merge all tsv files in a path into one data frame}
\usage{
multMerge(mypath)
}
\arguments{
\item{mypath}{the path containing files to load (all files will be loaded)}
}
\description{
This function opens all files in the given directory and merges
them into one dataframe. Files must meet the following requirement:\
     - Merge value in first column
     - Must have headers
     - Header value for data to be merged must be unique
}
\keyword{merge}

