% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{named.list}
\alias{named.list}
\title{Generates a list with names taken from variables used in its creation}
\usage{
named.list(...)
}
\arguments{
\item{...}{Elements to put in list}
}
\value{
A named list
}
\description{
Generates a list with names taken from variables used in its creation
}
\keyword{util}

